Functions: Implement our own which logic in have() function.
On distributions sbin is usually not in the users PATH and we fail
to detect dhcpd(3) and/or dnsmasq.

The have() function now uses our own which implmentation and
adds both /sbin and /usr/sbin when reading PATH. It will now return
the the full path when found or None if not.

Also update NetConf.py to use the updated have() function.