Catch exception when loading services
Adding a service may fail e.g. due to the corresponding dbus interface not being available when querying the state. Previously that led to all (following) services not being available in the menu. We now catch and print an exception and go on adding more services, so that only the failing items themselves will be missing.