RfKill: open /dev/rfkill r+b to avoid creating
Open /dev/rfkill in r+b mode rather than wb since the latter creates the
file if it does not exist. And we certainly can not create /dev/rfkill
as a regular file since this is never going to work and is going to
break a lot of other programs that assume /dev/rfkill is a valid device
node.

Fixes: https://github.com/blueman-project/blueman/issues/333