Put Python major version in shebang
This allows building for systems with both Python 2 and 3 without replacing the shebang. Replacing will only be necessary if env shall not be used e.g. as PyEnv might be present (see #435) or a specific minor version needs to be defined (i.e. there are multiple minor versions on the target system and blueman is not built for the newest one; we cannot use minor versions as they are not mandatory, see PEP 394).