Functions: Fix error formatter function e_()
We now split and always get the last item as this holds the error message.
Also call strip() on the string as it may contain a new line and the
MessageArea will show a "More" button.

This got broken with the GDBus migration of various parts in blueman as
GLib's exceptions do not match dbus-glib's.