Adapters: Create proper tabs again
For some reason max_width_chars does not make the label expand in Gtk3
when used as a page label in a GtkNotebook. Before we expanded the label
but this is hacky and ugly as it takes up he full width of notebook.

Instead we request a minimum label width which appears to be the only
way to get a reasonable width for the label.

It appears to be a know issue since at _least_ 2011, see bug 660649
https://bugzilla.gnome.org/show_bug.cgi?id=660649