BasePlugin: Add new on_delete functionality
This is useful for plugins that need to stop scripts and tear down
things that must always be done. For example kill running scripts like
we do in SerialManager.

We can not rely on __del__ to be run and this is most likely why
SerialManager uses the atexit module. We can howver use weakref to
run a function when the plugin is about tot be deleted.

Now we see the debug message "Deleting plugin instance %s" at shutdown
which never showed for me before.

But, this raises an interesting issue, we never see this message when
disabling a plugin. More work has to be done to properly unload plugins.