SerialManager: Run on_device_disconnect again when device is disconnects
This was broken with the introduction of generic (dis)connect. Also this
changed slightly as we now do not rely on the PluginManager to run
on_device_diconnect but run it when Connected property changed to False.

Import here is that the port is not active anymore at this point so
drop this condition otherwise we do nothing.