Add NetworkManager dun and pan connection classes
These classes handle the creation and activation of dun and pan
connections in NetworkManager.

They also store the connection uuid in gsettings so we can easilly
retrieve the connection in the future. I used org.blueman.gsmsettings
for both PANU and DUN as it is per device.

I chose to use the GObject-based client library as it makes things a lot
simpler than doing dbus cals directly.

The following commit(s) will move the NMPAN/NMDUNSupport plugins to the
new classes.