ManagerDeviceList: Use device.Icon as fallback
LE devices do not expose a class and instead we are provided with an
appearance. We already use appearance to create the caption but the icon
was still bluemans. We can however use device.Icon which usually has a
proper icon name.

This is similar to what was done in master, master however diverged too
much so had to re-implement it like this.