Fix time representation in NetUsage
This showed up in mypy 0.740 as ngettext is annotated to accept int but we provide float. The annotation is actually wrong as ngettext *does* accept float and just put it into round(), however rounding is not what we want here, at least not for hours, as e.g. 35 minutes would become 1 hour and 5 minutes.