Simplify NMConnectionBase
All usages of NMConnections pass reply and error handlers, so there is no point in having them optional. The reply_handler is either DbusService._handle_method_call.ok for the ConnectService or DisconnectService methods which just drops the values due to their signatures or AutoConnect._run.reply or RecentConns.on_item_activated.reply which do not care about the value (or not even take one since ab848697 😬), so there is no point in passing any.