Fix search in ManagerDeviceList
set_search_equal_func either must not receive user data or search_func must take it as an additional argument. It actually got added in #29 to the call but not to the method. I do not see how that could have worked but maybe something else changed that broke it along the way.