Handle unexpected DBus errors
The _call method is basically used in two ways:

* Make an asynchronous call, providing a reply and an error handler for when it's done.
* Make a call whose result we do not care about, providing no handlers.

In the latter case, we also do an asynchronous call which makes sense as the caller does not have to wait for it (might be an issue, though, as he might expect the call to already have an effect when the method returns). However, raising an exception when we get an error for that call does not make any sense as it will just get raised into the main loop (and get picked up by tools like Apport or ABRT), so best we can do here is just log the error and the call we got it for.
More involved solutions would be to either make the error_handler argument mandatory so that the caller has to think about errors showing up asynchronously or make a synchronous call if no error_handler is provided so that we're actually able to raise exceptions.