Handle added adapters in DiscvManager
I realized that this menu item was constantly missing if blueman-applet started with Bluetooth disabled. The reason is simply that the plugin checks the availability of an adapter when the D-Bus object manager becomes available or when the current adapter gets removed but not if an adapter gets added. It thus only shows the item if an adapter is available already when the object manager becomes available and only until it gets removed.