Install pixmaps in blueman datadir
Commit 800517a8c9b824f7c7fb90da584196016c19e771 (PR #132) moves
non-themed icons out of the XDG icon theme to a local directory.
The idea behind was definitely correct; IMHO the only issue was moving
the pixmaps to a public location (the 'pixmaps' subdirectory of the
system data directory) rather than to the blueman own data directory,
as they are private data of blueman rather than public pixmaps usable
by any application.

Hence, install the pixmaps to a subdirectory of the blueman data
directory, adjusting the paths that refer to this location.