Decouple battery indication from service presence
Up to now we try to read the battery percentage if and only if services are resolved and a battery service is present. That leads to two issues: On one hand there is the edge case where the interface is not provided anyway (due to some BlueZ internal issues), see #1708, on the other hand, the interface might still be there without the service, especially if the experimental battery provider API is used, see #1703.