Add Notificaton.set_message
The method can be used to replace the message especially of already shown notifications. _NotificationBubble.show() does keep track of the notification ID and uses the D-Bus protocol's replacing capabilities, so that we can just re-call it to replace an already shown notification. In case of _NotificationDialog it is sufficient to just set the property, no matter if the dialog is already shown or not.