Do not re-use dbusmenu item identifiers
Re-using identifiers e.g. for separator and standard menu items triggers bugs at least in gnome-shell-extension-appindicator (#1845).

As a workaround we can just use the toplevel items' priority values instead of sequential identifiers to get unique identifiers for all items.

To do that the MenuItem's priority value gets added as `id` to the MenuItemDict serialization and picked up by the IndicatorInterface implementations. Note that the semantics of the first argument ot the MenuItemActivator change accordingly.