Define relevant plugin types for BluemanApplet.Plugins
This is rather ugly, especially as it breaks the dynamic nature of the PluginManager and involves a cast, but it is definitely an improvement as it would have caught #1887 early.

There might be some generic way of defining this, like `__getattr__(self, key: Literal[_T]) -> _T` but I did not find a way to actually express something like "literal name for a typevar's value".