blueman-manager: Handle device removal events gracefully
As seen in https://bugzilla.redhat.com/show_bug.cgi?id=2148624, it looks like at least our fader can cause the device removal event to get handled multiple times, so that the tree_iter might actually be gone and the assertion causes an exception. If there is no tree_iter, there is nothing left that we could or should do, so we can just return.