Manager: Rework list sorting
Our list got very confusing with newly found device rows being inserted
in the list based on name or datetime added. When you have lot's of
devices it becomes a pain to find them so we need a better way.

We primarily now sort whether a device is paired. The logic behind this
is that these are the devices the user most likely is interested in. Next
we sort all connected devices below and the remaining ones at the bottom.