Remove check_bluetooth_status from blueman-manager
The function is completely unreliable regarding its option to turn Bluetooth on and regularly leads to closing the application while doing so (#1119 #2161). Also, the whole system is incosistent as while it avoids starting blueman-manager with Bluetooth powered off, you can absolutely run blueman-manager with Bluetooth powered off e.g. simply by turning it off while it's open. There is even a menu item for turnung Bluetooth off (and back on!) now.

Fun fact: Turning Bluetooth off while blueman-manager is running is actually meant to lead to the same dialog (and should have in 2.0), but #543 broke the signal handling: Since 2.1 `status` is set to a 1-tuple returned by `unpack` instead of a `bool` and the tuple is always `True`ish.

Note: blueman-adapters suffers from the same glitches (unreliable status check when turning Bluetooth on via the dialog, keeps running when Bluetooth gets turned off). That is not handled here.