Prepare 2.4.1
I plan to a add a note on icon themes to the release description on GitHub. Something like:

Note to maintainers: We found that the Adwaita Icon Theme dropped most non-symbolic icons in version 44, so that it can no longer be considered a generic fallback for GTK. blueman with just Adwaita Icon Theme will show a lot of missing icons. We are still puzzled by this change and try to figure out what to do. You can jump in at #2312. In the meantime, make sure that your packages depend on proper icons themes.