Handle incomplete StatusNotifierWatchers
kded6 provides an `org.kde.StatusNotifierWatcher` service without its usual `/StatusNotifierWatcher` object if Plasma Workspace's StatusNotifierWatcher module is registered but not loaded (https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/3301). We thus cannot just assume the whole service to be unavailable if the registration fails as we restart the tray application when the name appears.

This is a problem if the Plasma Workspace is around especially on i3 where no other application registers `org.kde.StatusNotifierWatcher`.

Closes #2311