Only show a singly battery notifcation
From feedback in #2317 it became pretty clear that battery data is a mess with devices providing levels like full, medium, low and devices providing fine-grained percent values, flapping values etc. and we do not want to go into power management applet business, so we only show a single battery connection (when the battery interface shows up or - if the interface is already present when the device connects - when the existing interface sends its first property update) and drop the reference to the connection notification to update so that further values do not lead to a notification. (In case people will ask for those notifications we might add them as a separate plugin that's disabled by default.)

Due to the great demand, I plan to backport this for 2.4.2, even though it's technically a regression.

Closes #2317