Fix audio profile menu item priorities
We build a (32 bit) ID for the menu item from its priority values, using the sub priority as the lower 8 bits. Pulseaudio card indexes are theoretically 32 bit and PipeWire re-assignes them with large gaps on re-connect so that it provides high 3-digit numbers in practice that do not fit the 8 bits.

Closes #2331