apps: parse log level with logging.getLevelName()
The logging.getLevelName() function returns the numeric representation of logging level.

The args.LEVEL is "warning" by default.
Just to be in line with the Python let's make it in upper case "WARNING".
But if previously an incorrect level name used e.g. CRIT then the WARNING fallback value used instead, but now it will fail.
This shouldn't be a big problem.