Sendto: Start discovery for new adapters
Start only after a second as adapters are not ready when discovery is
started right after it was added.

The callback expects a bool returned so _start_discovery now returns one
otherwise mypy complains it may return None. Would be nice if
GLib.timeout_add_once was available in pygobject :(.