Fix handling for new StatusNotifierWatcher (#2779)
blueman-tray restarts when a new StatusNotifierWatcher appears. This is especially relevant when blueman and a watcher like waybar get started in the session in no particular order. Unfortunately we have to use the `GLib.Error.message` and it looks like #2334 lacked the prefix; at least I cannot find any hint that this could have worked.

Fixes #2765